+++ packages/devtools_app/lib/src/flutter/scaffold.dart
@@ -38,7 +38,7 @@ class DevToolsScaffold extends StatefulWidget {
   static const Key fullWidthKey = Key('Full-width Scaffold');
 
   /// The width at or below which we treat the scaffold as narrow-width.
-  static const double narrowWidthThreshold = 1000.0;
+  static const double narrowWidthThreshold = 1060.0;
 
   /// The size that all actions on this widget are expected to have.
   static const double actionWidgetSize = 48.0;
@@ -243,8 +243,8 @@ class DevToolsScaffoldState extends State<DevToolsScaffold>
     }
 
     final appBar = AppBar(
-      // Turn off the appbar's back button on the web.
-      automaticallyImplyLeading: !kIsWeb,
+      // Turn off the appbar's back button.
+      automaticallyImplyLeading: false,
       title: title,
       actions: widget.actions,
       flexibleSpace: flexibleSpace,
